+++ packages/yourloops/test/unit/lib/data/data.api.test.ts
@@ -61,7 +61,7 @@ describe('Data API', () => {
       jest.spyOn(HttpService, 'get').mockRejectedValueOnce(Error('This error was thrown by a mock on purpose'))
       await expect(async () => {
         await DataApi.getPatientDataRange(patientId)
-      }).rejects.toThrowError('This error was thrown by a mock on purpose')
+      }).rejects.toThrow('This error was thrown by a mock on purpose')
     })
   })
 
@@ -114,7 +114,7 @@ describe('Data API', () => {
       jest.spyOn(HttpService, 'get').mockRejectedValueOnce(Error('This error was thrown by a mock on purpose'))
       await expect(async () => {
         await DataApi.getMessages({ userid: patientId } as IUser)
-      }).rejects.toThrowError('This error was thrown by a mock on purpose')
+      }).rejects.toThrow('This error was thrown by a mock on purpose')
     })
   })