+++ CHANGELOG.md
@@ -11,6 +11,8 @@ It is based on Tidepool Blip 1.27.
 - YLP-2227 Adjustment of the stats on carbs of the dashboard
 - YLP-2484 Remove level column in device change parameter history
 - YLP-2572 Prevent patient from adding the same caregiver to their list twice
+- YLP-2609 Moved patient monitoring alerts configuration to a dedicated view for HCP
+- YLP-2610 Direct link from Monitoring alerts card to "Target & alerts" view
 - YLP-2630 As a user, I would like to be notified with a banner when there are ongoing issues with yourloops
 
 ### Fixed
@@ -18,6 +20,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2627 "Last data update" column is not sorting properly
 - YLP-2642 Disabled Google Translate automated translation to avoid React crashes in notification page
 - YLP-2643 Disabled Google Translate automated translation to avoid React crashes in training page
+- YLP-2661 App crashes when patient does not have a pump
 - YLP-2662 App crashes for patient without enough data
 
 ### Engineering use